# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Order Me API"
        }
    ],
    "paths": {
...
        "/admin/order": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get list of order",
                "description": "Returns list of order",
                "operationId": "getOrders",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "code of order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "start date of order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "end date of order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit size ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page size ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "status of order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": " sort by query vd :-id,+id,+name,-name,-price,+price",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Create new order",
                "description": "Returns order data",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/order/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get order detail information",
                "description": "Returns order data",
                "operationId": "getOrderByIdAdmin",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDetailResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Order"
                ],
                "summary": "Update existing Order",
                "description": "Returns updated location data",
                "operationId": "updateOrderAdmin",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LocationResponse": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "address": {
                        "type": "string"
                    },
                    "distance": {
                        "type": "number"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "LogResponse": {
                "title": "Log Data",
                "description": "Log model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "change_by": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "OrderCreate": {
                "title": "Order Request",
                "description": "Order model",
                "properties": {
                    "phone": {
                        "type": "string"
                    },
                    "note": {
                        "type": "string"
                    },
                    "location_id": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "number"
                    },
                    "location_detail": {
                        "type": "string"
                    },
                    "dishes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderDetailRequest"
                        }
                    },
                    "payment_method": {
                        "type": "integer"
                    },
                    "price_sale": {
                        "type": "integer"
                    },
                    "price_none_sale": {
                        "type": "integer"
                    },
                    "coupon_id": {
                        "type": "integer"
                    },
                    "payment_status": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "OrderDetailRequest": {
                "title": "Order Detail Request",
                "description": "Order model",
                "properties": {
                    "dish_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "price_sale": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "OrderDetailResponse": {
                "title": "Order Detail Response",
                "description": "Order model",
                "properties": {
                    "dishes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderDishItem"
                        }
                    },
                    "logs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LogResponse"
                        }
                    },
                    "total": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "OrderDishItem": {
                "title": "Order Dish Item",
                "description": "Order model",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "image": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "price": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "OrderResponse": {
                "title": "Order Request",
                "description": "Order model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "code": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "total": {
                        "type": "string"
                    },
                    "note": {
                        "type": "string"
                    },
                    "location": {
                        "$ref": "#/components/schemas/LocationResponse"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "vnp_url": {
                        "type": "string"
                    },
                    "location_detail": {
                        "type": "string"
                    },
                    "price_sale": {
                        "type": "integer"
                    },
                    "price_none_sale": {
                        "type": "integer"
                    },
                    "coupon_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "OrderUpdate": {
                "title": "Order Request",
                "description": "Order model",
                "properties": {
                    "status": {
                        "type": "integer"
                    },
                    "phone": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UserResponse": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
...
}